[storage] move codec for NodeKey from schema to jellyfish merkle
