+++ storage/jellyfish_merkle/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 bincode = "1.1.1"
+byteorder = "1.3.2"
 num-derive = "0.2"
 num-traits = "0.2"
 proptest = "0.9.2"
